requirements.txt
Flask==3.0.3
pandas==2.2.2
kagglehub==0.3.3
google-cloud-bigquery==3.25.0
gunicorn==22.0.0

